requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.30.5
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==2.0.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
youtube-transcript-api==0.6.2
python-dateutil==2.9.0
pytest==8.2.2
